refactor: Extract user command functionality to separate module
